Fix APA102 bug causing the last LED to not turn on
The APA102 datasheets say that 0xFFFFFFFF should be sent as the end
frame of a command to control a string of LEDs. That doesn't work if
the numbers of LEDs in the string is greater than the number of LEDs
addressed in the command. The 0xFFFFFFFF end frame will be interpreted
as a command to the subsequent LED to shine white with max brightness.
That LED will turn on the next time a start frame (0x00000000) is sent.

Without this patch, the code to send the end frame (in endBoundary())
is commented out. That solves the problem with the max white LED, but
unfortunately it means that the last LED of a command will not be turned
on. This bug can be reproduce by sending a command to turn on two LEDs
and then call FastLED.show(). The second LED will not turn on until the
next time FastLED.show() is called.

The fix is to send 0x00000000, the start frame, instead of 0xFFFFFFFF.
That turns on the last LED and seems to reset the internal state of all
LEDs in the LED strip. After the reset, subsequent LED string commands
work as before.